+++ packages/core/src/js/feedback/defaults.ts
@@ -1,6 +1,6 @@
 import { Alert } from 'react-native';
 
-import type { FeedbackButtonProps, FeedbackFormProps } from './FeedbackForm.types';
+import type { FeedbackButtonProps, FeedbackWidgetProps } from './FeedbackWidget.types';
 
 const FORM_TITLE = 'Report a Bug';
 const NAME_PLACEHOLDER = 'Your Name';
@@ -21,7 +21,7 @@ const ADD_SCREENSHOT_LABEL = 'Add a screenshot';
 const REMOVE_SCREENSHOT_LABEL = 'Remove screenshot';
 const GENERIC_ERROR_TEXT = 'Unable to send feedback due to an unexpected error.';
 
-export const defaultConfiguration: Partial<FeedbackFormProps> = {
+export const defaultConfiguration: Partial<FeedbackWidgetProps> = {
   // FeedbackCallbacks
   onFormOpen: () => {
     // Does nothing by default
@@ -34,7 +34,7 @@ export const defaultConfiguration: Partial<FeedbackFormProps> = {
       );
     }
   },
-  onAddScreenshot: (_: (filename: string, data: Uint8Array) => void) => {
+  onAddScreenshot: (_: (uri: string) => void) => {
     if (__DEV__) {
       Alert.alert('Development note', 'onAddScreenshot callback is not implemented.');
     }